backend_api/pip_requirements.txt
psycopg2==2.7.6.1
django==2.1.4
djangorestframework==3.9.0
pylama==7.6.6
